robots/requirements.txt
playwright>=1.40.0
mysql-connector-python>=8.2.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
